seperate servers and clients
